+++ project/src/com/test/AdminMenu.java
@@ -276,12 +276,13 @@ public class AdminMenu {
     //단어관리 수정/추가 기능
     word: while(true) {
       System.out.println();
-      System.out.print("[1. 단어 추가]   [2. 단어 삭제]   [3. 단어 목록]   [8. 뒤로가기]   \n >>> ");
+      System.out.print("[1. 단어 추가]   [2. 단어 삭제]   [3. 단어 목록]   [4. 단어 검색]   [8. 뒤로가기]   \n >>> ");
       int menu = Integer.parseInt(sc.nextLine());
       switch(menu) {
         case 1: addWord(); break;
         case 2: delWord(); break;
         case 3: printWord(); break;
+        case 4: searchWord(); break;
         case 8: break word;
         default: System.out.println("메뉴 번호 확인"); continue;
       }
@@ -290,14 +291,55 @@ public class AdminMenu {
   }
 
   public void addWord() {
+    JoinMember jm = new JoinMember();
     try {
       System.out.println("\n[단어 추가]");
-      System.out.print("\n  level>>> ");
-      int wlevel = Integer.parseInt(sc.nextLine());
-      System.out.print("  영단어>>> ");
-      String weng = sc.nextLine();
-      System.out.print("  단어 뜻>>> ");
-      String wkor = sc.nextLine();
+      int wlevel = 0;
+      String weng = "추가할 영단어";
+      String wkor = "추가할 단어뜻";
+      while(true) {
+        System.out.print("\n  level>>> ");
+        try{wlevel = Integer.parseInt(sc.nextLine());
+        if(wlevel<1 || wlevel>3) {
+          System.out.println("1~3 사이의 숫자를 입력해주세요.\n");
+          continue;
+        }
+        }catch(Exception ex) { System.out.println("숫자를 입력해주세요.\n"); continue;}
+        break;
+      }//while end
+
+      eng:
+        while(true) {
+          while(true) {
+            System.out.print("  영단어>>> ");
+            weng = sc.nextLine();
+            if(jm.stringCheck(weng) || weng.matches(".*[0-9ㄱ-ㅎㅏ-ㅣ가-힣]+.*"  )) {
+              System.out.println("영어로 입력해주세요.");
+              continue;
+            }
+            break;
+          }
+          sql = "select * from word where wordLevel = " + wlevel;
+          RS = ST.executeQuery(sql);
+          while(RS.next() == true) {
+            String eng = RS.getString("eng");
+            if(weng.equals(eng)) {
+              System.out.println("이미 등록된 단어입니다.");
+              continue eng;
+            }//if end
+          }//while end
+          break;
+        }//while end
+
+      while(true) {
+        System.out.print("  단어 뜻>>> ");
+        wkor = sc.nextLine();
+        if(jm.stringCheck(wkor) || wkor.matches("^[a-zA-Z0-9]*$")) {
+          System.out.println("한글로 입력해주세요.");
+          continue;
+        }
+        break;
+      }
 
       sql = "insert into word(wordNum, wordLevel, eng, kor) values( word_seq.nextval, " + wlevel + ", '" + weng + "', '" + wkor + "')";
       int check = ST.executeUpdate(sql);
@@ -315,20 +357,32 @@ public class AdminMenu {
   }
 
   public void delWord() {
+    String eng = "삭제할 단어";
     try {
       System.out.println("\n[단어 삭제]");
-      System.out.print("\n  삭제할 단어 입력>>> ");
+      System.out.print("\n  삭제할 단어 입력 (영어)>>> ");
       String weng = sc.nextLine();
-      sql = "delete from word where eng = '" + weng + "'";
-      ST.executeUpdate(sql);
-      System.out.println("\n단어 삭제 완료");
-      System.out.print("\n  단어 목록 출력? (y/N)>>> ");
-      String print = sc.nextLine();
-      if (print.equals("y")) {
-        printWord();
-      } else {
-        return;
-      }
+
+      sql = "select * from word";
+      RS = ST.executeQuery(sql);
+      while(RS.next() == true) {
+        eng = RS.getString("eng");
+        if(weng.equals(eng)) {
+          sql = "delete from word where eng = '" + weng + "'";
+          ST.executeUpdate(sql);
+          System.out.println("\n  단어 삭제 완료");
+          System.out.print("\n  단어 목록 출력? (y/N)>>> ");
+          String print = sc.nextLine();
+          if (print.equals("y")) {
+            printWord();
+          } else {
+            return;
+          }//if end
+        }//if end
+      }//while end
+      if(!weng.equals(eng)) {
+        System.out.println("\n  삭제할 수 없는 단어입니다.");
+      }//if end
     } catch(Exception e) {System.out.println("error delword:" + e);}
   }
 
@@ -338,6 +392,7 @@ public class AdminMenu {
       sql = "select wordLevel, eng, kor from word order by wordLevel";
       RS = ST.executeQuery(sql);
       System.out.println("\n영단어\t단어뜻");
+      System.out.println("------------------------------------------------------------");
       while(RS.next() == true) {
         String peng = RS.getString("eng");
         String pkor = RS.getString("kor");
@@ -346,11 +401,40 @@ public class AdminMenu {
     } catch(Exception e) {System.out.println("error printWord:" + e);}
   }
 
+  public void searchWord() {
+    String type = "영어,한글";
+    int level = 0;
+    String eng ="영단어", kor = "뜻", sword = "검색단어";
+    try {
+      while(true) {
+        System.out.println("검색할 단어를 입력해주세요.");
+        sword = sc.nextLine();
+        if(sword.matches("^[a-zA-Z]*$")) {type = "eng";
+        }else if(sword.matches(".*[ㄱ-ㅎㅏ-ㅣ가-힣]+.*")) {type = "kor";
+        }else {System.out.println("영어나 한글로 입력해주세요.\n");continue;}
+        break;
+      }
+      sql = "select * from word where " + type + " like '%" + sword + "%'";
+      RS = ST.executeQuery(sql);
+      System.out.println("레벨\t영어\t\t\t뜻\n");
+      System.out.println("------------------------------------------------------------");
+      while(RS.next() == true) {
+        level = RS.getInt("wordLevel");
+        eng = RS.getString("eng");
+        kor = RS.getString("kor");
+        System.out.printf("%d\t%s\t\t%s\n\n", level, eng, kor);
+      }//while end
+      if(eng.equals("영단어")) {
+        System.out.println("\n목록에 없는 단어입니다.");
+      }
+    }catch(Exception ex) { }
+  }//searchWord end
+
   public void notice() {
     try {
       Loop: while(true) {
         System.out.println("\n[공지 관리]");
-        System.out.print("[1. 공지 등록]   [2. 공지 목록]   [8. 뒤로가기]\n >>> ");
+        System.out.print("[1. 공지 등록]   [2. 공지 목록]   [3. 공지 삭제]   [4. 공지 상세조회]   [8. 뒤로가기]\n >>> ");
         String cmd = sc.nextLine();
         switch(cmd) {
           case "1":
@@ -359,6 +443,11 @@ public class AdminMenu {
             String nTitle = sc.nextLine();
             System.out.print("  공지 내용 >>> ");
             String nContent = sc.nextLine();
+            while(nContent.length()>100) {
+              System.out.println("\n100자이내로 작성해주세요");
+              System.out.print("\n 공지 내용 >>> ");
+              nContent = sc.nextLine();
+              continue; } 
             String msg = "insert into notice(code, title, content) values(notice_seq.nextval, '"+nTitle+"', '"+nContent+"')";
             ST.executeUpdate(msg);
             System.out.println("제목: " + nTitle);
@@ -366,15 +455,22 @@ public class AdminMenu {
             System.out.println("공지 등록 완료");
             break;
           case "2":
-            msg = "select code, title, content from notice order by code";
+            msg = "select * from notice order by code asc";
             RS = ST.executeQuery(msg);
-            System.out.println("\nNo. \t Title \t\t\t\t\t Content");
+            System.out.println("\nNo.\t  Date  \t Title");
+            System.out.println("------------------------------------------------------------");
             while(RS.next() == true) {
               int pcode = RS.getInt("code");
+              Date pdate = RS.getDate("cdate");
               String ptitle = RS.getString("title");
-              String pcontent = RS.getString("content");
-              System.out.println(pcode + "\t" + ptitle + " \t\t\t\t\t " + pcontent);
+              System.out.println(pcode + "\t  " + pdate + "  \t " + ptitle);
             }
+            break;
+          case "3":
+            delNotice();
+            break;
+          case "4":
+            detailsNotice();
             break;
           case "8":
             break Loop;
@@ -385,4 +481,73 @@ public class AdminMenu {
       }
     } catch(Exception e) {System.out.println("error notice:" + e);}
   }
+
+  public void delNotice() {
+    String msg;
+    int delNum = 0;
+    while(true) {
+      System.out.println("삭제할 게시물 번호를 입력해주세요.");
+      try {delNum = Integer.parseInt(sc.nextLine()); 
+      msg = "select * from notice where code = " + delNum;
+      RS = ST.executeQuery(msg);
+      if (RS.next() == true) {
+        int pcode = RS.getInt("code");
+        String ptitle = RS.getString("title");
+        String pcontent = RS.getString("content");
+        Date pdate = RS.getDate("cdate");
+        System.out.println("\nNo.      :\t" + pcode);
+        System.out.println("Date     :\t" + pdate);
+        System.out.println("Title    :\t " + ptitle);
+        System.out.print("Content  :\t" );
+        AccountInfo ai = new AccountInfo();
+        ai.printcomnt(pcontent);
+      }else {
+        System.out.println("공지번호를 다시 확인하십시오.\n");
+        break;
+      }//if end
+
+      System.out.print("\n정말 삭제하시겠습니까? (y/N)>>> ");
+      String a = sc.nextLine();
+      if (a.equals("y")) {
+        msg = "delete from Notice where code = " + delNum;
+        ST.executeUpdate(msg);
+        System.out.println("공지 삭제가 완료되었습니다.");
+        System.out.println();
+        break;
+      } else {System.out.println("공지 삭제를 취소하셨습니다.");}
+      }catch(Exception ex) {System.out.println("숫자를 입력해주세요.\n");continue;}
+      break;
+    }//while end
+  }
+  public void detailsNotice(){
+    int pcode = 0;
+    int num = 0;//사용자입력값
+    String ptitle = "제목", pcontent = "내용";
+    Date pdate;
+    while(true) {
+      try {
+        dbConnect();
+        System.out.println("상세조회할 공지 번호를 입력해주세요.");
+        num = Integer.parseInt(sc.nextLine());
+        sql = "select * from notice where code = " + num;
+        RS = ST.executeQuery(sql);
+        if (RS.next() == true) {
+          pcode = RS.getInt("code");
+          ptitle = RS.getString("title");
+          pcontent = RS.getString("content");
+          pdate = RS.getDate("cdate");
+          System.out.println("\nNo.      :\t" + pcode);
+          System.out.println("Date     :\t" + pdate);
+          System.out.println("Title    :\t " + ptitle);
+          System.out.print("Content  :\t" );
+          AccountInfo ai = new AccountInfo();
+          ai.printcomnt(pcontent);
+        }else {
+          System.out.println("공지번호를 다시 확인하십시오.\n");
+          break;
+        }//if end
+      }catch(Exception ex) {System.out.println("\n공지번호로 입력하셔야합니다."); }
+      break;
+    }//while end
+  }//noticeDetail end
 }
